notes_app/requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.12
mysqlclient>=2.0
django-cors-headers>=3.10
djangorestframework-simplejwt>=5.2
Pillow>=9.0
pytest-django>=4.5
pytest>=7.0
